requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
prophet==1.1.5
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
kafka-python==2.0.2
mlflow==2.10.2
prometheus-client==0.19.0
psycopg2-binary==2.9.9
influxdb-client==1.40.0
pytest==8.0.0
python-dotenv==1.0.1
